+++ proto/terra/gov/v2lunc1/gov.proto
@@ -9,7 +9,7 @@ import "google/protobuf/duration.proto";
 import "cosmos_proto/cosmos.proto";
 import "amino/amino.proto";
 
-option go_package = "github.com/classic-terra/core/v3/custom/gov/types";
+option go_package = "github.com/classic-terra/core/v3/custom/gov/types/v2lunc1";
 
 // Params defines the parameters for the x/gov module.
 //
