+++ proto/terra/gov/v2lunc1/tx.proto
@@ -10,7 +10,7 @@ import "cosmos_proto/cosmos.proto";
 import "cosmos/msg/v1/msg.proto";
 import "amino/amino.proto";
 
-option go_package = "github.com/classic-terra/core/v3/custom/gov/types";
+option go_package = "github.com/classic-terra/core/v3/custom/gov/types/v2lunc1";
 
 // Msg defines the gov Msg service.
 service Msg {
